ci: fix webdriverio test using wrong chromedriver
The webdriverio CI job was failing with the error message:

    session not created: This version of ChromeDriver only supports Chrome version 109
    Current browser version is 111.0.5563.64 with binary path /usr/bin/google-chrome

Despite the job running the following steps:

    - browser-tools/install-chrome
    - browser-tools/install-chromedriver
    - run: npm run coverage --prefix=packages/webdriverio

And browser-tools/install-chromedriver automatically installing the
right version of chromedriver based on the previously installed chrome
version.[1]

This was because browser-tools/install-chromedriver installed the
chromedriver under /usr/local/bin/chromedriver while
packages/webdriverio/src/test.ts was using the chromedriver from the
chromedriver NPM package (which is currently set to ^109.0.0 in
package.json).

This commit changes the webdriverio test code to instead use the
system-wide installed chromedriver (that has been installed via
browser-tools/install-chromedriver) in a CI environment.

[1]: https://circleci.com/developer/orbs/orb/circleci/browser-tools#commands-install-chromedriver