chore: replace substr for browser check (#1202)
Saw this file was using a deprecated API to check for the existence of
one string inside of another. Converted it to the newer API for this
purpose.

Extra unit tests added thanks to @dbjorge for catching a condition the
previous tests were not accounting for.

QA shouldn't need to do any manual testing. This is well covered by the
unit test which passes. If anyone wants to validate further, run the
command with a few browser options and provide some un-finished
versions. (Un-finished as in `ch` for `chrome`)

No QA needed
Refs: #1204