ci(tests): disable fail-fast (#1227)
This patch disables the `fail-fast` strategy. That way if certain parts
of the matrix fail they don't prevent completion of other parts that do
succeed but haven't had enough time to report.

It also drops a one-time log onto Puppeteer (since it is giving us the
most trouble) so we can verify what Chrome version it is using for us to
try and replicate problems locally.

I saw the husky log for the script that would break in the next major.
Removed the lines in question while here.

No QA Required