chore: mark test packages as private (#1241)
This patch marks the remaining two test packages as `private`. It also
patches the main workspaces configuration to include all the test
packages.

Previously the wdio test was published by accident. It was marked as a
workspace but not as private. The `release` test package was not marked
as private, but was not included in the workspaces. Therefore it was not
published by accident alone.

Now all packages are available as a workspace, while keeping things
private that should be.

Fixes: #1238